project order issue -after deleting a project, fixed

+++ app/Http/Controllers/ProjectsController.php
@@ -14,6 +14,8 @@ use App\Project;
 use App\Sector;
 use App\TargetGroup;
 use App\User;
+use DB;
+
 use Auth;
 use Illuminate\Http\Request;
 use Maatwebsite\Excel\Facades\Excel;
@@ -369,7 +371,15 @@ class ProjectsController extends Controller
     public function destroy($id)
     {
         $project = Project::findOrFail($id);
+    
         $project->delete();
+
+        DB::table('projects')
+    ->where('user_id', $project->user_id)
+    ->where('project_order', '>', $project->project_order)
+    ->decrement('project_order', 1);
+
+
         return redirect('/projects')->with('error', 'Project deleted');
 
     }
